License project under Creative Commons Zero v1.0 Universal
It is important to define the license under which a software project is released in order to permit its use.

Previously, this project did not contain any information about licensing.

This project is not truly a library, but rather just a collection of sketches packaged as a library in order to allow
distribution via Library Manager, but without any library source code. So it is those sketches that are the significant
content of the project, for which licensing must be defined.

It is standard practice for Arduino to release sketches into the public domain. The appropriate formal license in this
case is Creative Commons Zero v1.0 Universal.

A standardized Creative Commons Zero v1.0 Universal license file is hereby added to the project. In addition to
providing this important information to the project users, this will also allow automated license type detection (e.g.,
by the Licensee Gem used by the GitHub website).